rc4 completed but don't have input text by file
